adds dynamic emoji loading to AllProducts
